Add caption and topic fields to upload page
This commit adds a caption and topic field to the upload page.

The caption field is a text input that allows the user to add a caption to their video.

The topic field is a select input that allows the user to select a topic for their video.

The topic field is populated with the topics from the constants file.